ai_server/requirements.txt
Flask==3.0.0
gunicorn==22.0.0
python-dotenv==1.0.1

numpy==1.26.4          
pandas==2.2.2
scikit-learn==1.4.2
networkx==3.2.1

torch==2.4.1
xgboost==2.1.1
lightgbm==4.5.0
sentence-transformers==2.7.0

faiss-cpu==1.9.0        

redis==5.0.7
pymysql==1.1.1
SQLAlchemy==2.0.35

pytest==8.2.1
black==24.8.0
flake8==7.1.1
setuptools>=69.0.0
wheel>=0.42.0
